Initial support for Cloudflare protected streaming sites
This addition makes it possible to scrape Cloudflare protected websites. It keeps the current API, but adds an overload for POST requests. The performance, after the initial Cloudflare handshake should be comparable to the previous iteration, but the first connection to any protected server will take at least 5 seconds.

This commit also includes a basic implementation for KissAnime searching.