The big Hummingbird update. The API is linked in and everything but there's no scrobbling as of yet with the hummingbird integration limited to just logging in and seeing your sweet avatar and cover in the sidebar. It's probably riddled with bugs :)  ...I should comment at least a little bit ;-;
The settings fragment has gotten some cleanup with less insane title handling.

The way fragments are persisted / the navigation tree has been completely reworked.

Drawer peeks behind the status bar. In the same vein, user stock images are included so the drawer doesn't look like garbage if the user doesn't have a hummingbird (the images will probably be changed).

There's a history tab in Hummingbird Settings that does absolutely nothing as of yet but I can't be bothered to remove it since it will eventually be used for monitoring what updates have been made by the app - it will basically be the user's feed  (though only library updates).

Is that everything? Yay.